⭐️ Solve day 3 in Go
I've written Go before, but it's been a few years. While the docs are
not the most aesthetic, It was quite easy to find the packages and
functions I needed to get the job done.

I felt like a bad gopher not checking the `err` values of functions...
but for these purposes, `panic`'ing would have been just fine.

I still don't know when to distinguish between

```
var something = []string{}
something := make([]string, 0)
```