🎁 Solve Day 4 with JavaScript
Everyone's favorite language. The big hang up on this one was parsing
dates. During my first pass, I split appart all the components of the
date, and was doing `new Date(year, month, day, hour, minute)`...  and I
learned that the month really needed to be month minus one. After that,
timezones were also an issue... but passing in the milliseconds and Z
got that all sorted.

I wish there was a way to natively `mapKeys` on an object since that
seemed to be the bulk of problem solving for this particular problem.
`Object.keys` to construct a new object got me close enough.